web-app-frameworks/streamlit-chat-app/requirements.txt
streamlit==1.29.0
websocket-client==1.7.0